+++ icicle.cabal
@@ -122,12 +122,6 @@ library
 
                        Icicle.Dictionary
                        Icicle.Dictionary.Data
-                       Icicle.Dictionary.Import
-                       Icicle.Dictionary.Parse
-                       Icicle.Dictionary.Parse.Types
-                       Icicle.Dictionary.Parse.Toml
-                       Icicle.Dictionary.Parse.Prisms
-                       Icicle.Dictionary.Parse.TomlDictionary
                        Icicle.Dictionary.Demographics
 
                        Icicle.Encoding
@@ -169,6 +163,15 @@ library
                        Icicle.Source.ToCore.Fold
                        Icicle.Source.ToCore.ToCore
                        Icicle.Source.ToCore.Prim
+
+                       Icicle.Storage.Dictionary.TextV1
+                       Icicle.Storage.Dictionary.Toml.Import
+                       Icicle.Storage.Dictionary.Toml.Types
+                       Icicle.Storage.Dictionary.Toml.Toml
+                       Icicle.Storage.Dictionary.Toml.Prisms
+                       Icicle.Storage.Dictionary.Toml.TomlDictionary
+
+                       Icicle.Storage.Encoding
 
                        Icicle.Repl
