+++ src/Icicle/Storage/Dictionary/TextV1.hs
@@ -1,12 +1,9 @@
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE LambdaCase       #-}
-module Icicle.Dictionary.Parse (
+module Icicle.Storage.Dictionary.TextV1 (
     parseDictionaryLineV1
   , writeDictionaryLineV1
-  , parsePrimitiveEncoding
-  , parseEncoding
-  , prettyConcrete
   ) where
 
 import           Icicle.Data
@@ -18,6 +15,8 @@ import           Data.Attoparsec.Text
 
 import           Data.Either.Combinators
 import           Data.Text hiding (takeWhile)
+
+import           Icicle.Storage.Encoding
 
 field :: Parser Text
 field = append <$> takeWhile (not . isDelimOrEscape) <*> (concat <$> many (cons <$> escaped <*> field)) <?> "field"
@@ -37,27 +36,6 @@ parseIcicleDictionaryV1 = do
     where
       p = char '|'
 
-parsePrimitiveEncoding :: Parser Encoding
-parsePrimitiveEncoding =
-           StringEncoding  <$ string "string"
-       <|> IntEncoding     <$ string "int"
-       <|> IntEncoding     <$ string "long" -- Todo, change this once Longs are a thing
-       <|> DoubleEncoding  <$ string "double"
-       <|> DateEncoding    <$ string "date"
-       <|> BooleanEncoding <$ string "boolean"
-
-parseEncoding :: Parser Encoding
-parseEncoding = parsePrimitiveEncoding
-       <|> ListEncoding    <$ char '[' <*> parseEncoding <* char ']'
-       <|> StructEncoding  <$ char '(' <*> (structField `sepBy` char ',') <* char ')'
-  where
-    structField = do
-      n <- takeWhile (/= ':')
-      _ <- char ':'
-      e <- parsePrimitiveEncoding
-      o <- Optional <$ char '*' <|> pure Mandatory
-      pure $ StructField o (Attribute n) e
-
 parseDictionaryLineV1 :: Text -> Either ParseError DictionaryEntry
 parseDictionaryLineV1 s =
   mapLeft (ParseError . pack) $ parseOnly parseIcicleDictionaryV1 s
@@ -67,17 +45,3 @@ writeDictionaryLineV1 (DictionaryEntry (Attribute a) (ConcreteDefinition e)) =
   a <> "|" <> prettyConcrete e
 
 writeDictionaryLineV1 (DictionaryEntry _ (VirtualDefinition _)) = "Virtual features not supported in V1"
-
-prettyConcrete :: Encoding -> Text
-prettyConcrete = \case
-  StringEncoding   -> "string"
-  IntEncoding      -> "int"
-  DoubleEncoding   -> "double"
-  DateEncoding     -> "date"
-  BooleanEncoding  -> "boolean"
-  ListEncoding le  -> "[" <> prettyConcrete le <> "]"
-  StructEncoding s -> "(" <> intercalate "," (prettyStructField <$> s) <> ")"
-
-prettyStructField :: StructField -> Text
-prettyStructField (StructField Mandatory (Attribute n) e) = n <> ":" <> prettyConcrete e
-prettyStructField (StructField Optional (Attribute n) e) = n <> ":" <> prettyConcrete e <> "*"
